add default xarray container and update version
