+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "deklare"
-version = "0.1.3"
+version = "0.1.4"
 description = "A data processing framework for data exploration"
 authors = [
     { name = "Matthias Meyer", email = "contact@niowniow.net" },
@@ -16,6 +16,7 @@ dependencies = [
     "pystac>=1.11.0",
     "shapely>=2.0.6",
     "tqdm>=4.67.1",
+    "xarray>=2025.6.1",
 ]
 readme = "README.md"
 requires-python = ">= 3.10"
