+++ src/deklare/persist.py
@@ -13,220 +13,28 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License."""
 
-import io
 import json
-import warnings
+import fsspec
+
 from copy import copy, deepcopy
 from pathlib import Path
 from threading import Lock
-from typing import Callable, TypeVar, Type, NamedTuple, Self, IO
-from abc import ABC, abstractmethod
+from typing import Callable, Type
 
-import pandas as pd
 from cachetools import LRUCache
-from compress_pickle import dump, load
 
 # TODO: can we implement our own hash function for deskriptors to reduce dependency on dask?
 from dask.base import tokenize
 
 import pystac
-from shapely.geometry import Polygon, mapping
 
 from .core import task
 from .utils import (
     NodeFailedException,
     dict_update,
-    exclusive_indexing,
-    indexers_to_slices,
+    get_segments,
 )
-
-import fsspec
-
-from .utils import get_segments
-
-class MediaDescription(NamedTuple):
-    media_type: str
-    description: str
-
-class StacIO(pystac.StacIO):
-    def __init__(self, store):
-        self.store = store
-
-    def read_text(self, source: pystac.utils.HREF, *args, **kwargs) -> str:
-        """Reads the data at `source` stored in the store
-
-        Args:
-            source : The source to read from.
-
-        Returns:
-            str: The text contained in the file at the location specified by the uri.
-        """
-        str_src = str(source)
-        if str_src.startswith('/') and len(str_src) > 1:
-            str_src = str_src[1:]
-        return self.store[str_src].decode()
-
-    def write_text(self, dest: pystac.utils.HREF, txt: str, *args, **kwargs) -> None:
-        """writes the data of `txt` into the store to `dest`
-
-        Args:
-            dest : The destination to write to.
-            txt : The text to write.
-        """
-        str_dest = str(dest)
-        if str_dest.startswith('/') and len(str_dest) > 1:
-            str_dest = str_dest[1:]
-
-        self.store[str_dest] = txt.encode()
-
-    def gen_stac_item_kwargs(self, deskriptor: dict, item_metadata: dict) -> dict:
-        """generates metadata for the stac item for a deskriptor
-
-        Args:
-            deskriptor (dict): deskriptor describing
-            item_metadata (dict): _description_
-
-        Raises:
-            RuntimeError: _description_
-
-        Returns:
-            dict: _description_
-        """
-        if (
-            "longitude" not in deskriptor or
-            "start" not in deskriptor["longitude"] or
-            "end" not in deskriptor["longitude"] or
-            "latitude" not in deskriptor or
-            "start" not in deskriptor["latitude"] or
-            "end" not in deskriptor["latitude"] or
-            "time" not in deskriptor or
-            "variable" not in deskriptor
-        ):
-            raise RuntimeError("Given deskriptor does not match required metadata format")
-
-        id = deskriptor["deskriptor_hash"]
-        bbox = [
-            deskriptor["longitude"]["start"],
-            deskriptor["latitude"]["end"],
-            deskriptor["longitude"]["end"],
-            deskriptor["latitude"]["start"],
-        ]
-        footprint = mapping(
-            Polygon(
-                [
-                    [bbox[0], bbox[1]],  # lower left corner
-                    [bbox[0], bbox[3]],  # upper left corner
-                    [bbox[2], bbox[3]],  # upper right corner
-                    [bbox[2], bbox[1]],  # lower right corner
-                    [bbox[0], bbox[1]],  # lower left corner
-                ]
-            )
-        )
-        start_time = deskriptor["time"]["start"].to_pydatetime()
-        end_time = deskriptor["time"]["end"].to_pydatetime()
-        variables = deskriptor["variable"]
-
-        kwargs = {
-            "id": id,
-            "geometry": footprint,
-            "bbox": bbox,
-            "datetime": None,
-            "start_datetime": start_time,
-            "end_datetime": end_time,
-            "properties": {
-                "description": self._gen_description(deskriptor),
-                "variables": variables,
-            },
-        }
-
-        blocked_keys = kwargs.keys()
-        for k, v in item_metadata.items():
-            if k not in blocked_keys:
-                kwargs[k] = v
-            elif k == "properties" and isinstance(v, dict):
-                for k_p, v_p in v.items():
-                    if k_p != "variables":
-                        kwargs[k][k_p] = v_p
-
-        return kwargs
-
-    def _gen_description(self, deskriptor) -> str:
-        """generate human readable description for STAC Item of chunk
-
-        Returns:
-            str: STAC Item description
-        """
-        start_time = deskriptor["time"]["start"].isoformat()
-        end_time = deskriptor["time"]["end"].isoformat()
-        variable_string = ", ".join(deskriptor["variable"])
-        latitude_string = f"{deskriptor['latitude']['start']} to {deskriptor['latitude']['end']} latitude"
-        longitude_string = f"{deskriptor['longitude']['start']} to {deskriptor['longitude']['end']} longitude"
-
-        description = (
-            f"This chunk contains data for the variable(s) {variable_string}, "
-            f"collected from {start_time} to {end_time}, "
-            f"covering the geographic region defined by {latitude_string} and {longitude_string}"
-        )
-
-        return description
-
-
-class DataContainer(ABC):
-    @abstractmethod
-    def write(self, file: IO) -> None:
-        """Takes the data and writes it to the file
-
-        Args:
-            file (IO): file-like object to save data from itself too
-        """
-        pass
-
-    @classmethod
-    @abstractmethod
-    def read(cls, file: IO) -> Self:
-        """Reads the data in file and creates new data container from it
-
-        Args:
-            file (IO): file-like object containing saved data
-
-        Returns:
-            DataContainer: Container containing data saved in file
-        """
-        pass
-
-    @abstractmethod
-    def get_stac_metadata(self) -> dict | None:
-        """optionally returns STAC metadata for the contained data
-
-        returns:
-            dict | None: dict containing STAC metadata or None for no metadata
-        """
-        pass
-
-    @classmethod
-    @abstractmethod
-    def merge(cls, *elements: Self) -> Self:
-        """merge the elements into a single DataContainer instance
-        if no elements are passed return an empty container, if only one is passed acts as the identity function
-
-        Args:
-            elements (DataContainer): list of data containers to merge
-
-        Returns:
-            DataContainer: DataContainer resulting from merging elements
-        """
-        pass
-
-    @abstractmethod
-    def file_info(self) -> MediaDescription:
-        """return media type and text describtion of file saved in `write` function
-        media type should be registered in https://www.iana.org/assignments/media-types/media-types.xhtml e.g. `image/tiff`
-        description should be human readable with information needed to read file
-
-        Returns:
-            MediaDescription: (media type, description)
-        """
-        pass
+from .data_io import StacIO, DataContainer, MediaDescription
 
 
 @task()
@@ -374,7 +182,6 @@ class Persister:
             except Exception as e:
                 print("Error during Persister", repr(e))
 
-
             return data
         else:
             raise NodeFailedException("A bug in Persister. Please report.")
@@ -417,7 +224,9 @@ class Persister:
 
         return deskriptor_hash
 
-    def _save_metadata(self, deskriptor: dict, item_metadata: dict, file_info: MediaDescription) -> None:
+    def _save_metadata(
+        self, deskriptor: dict, item_metadata: dict, file_info: MediaDescription
+    ) -> None:
         """saves metadata for given chunk using STAC (https://stacspec.org/)
 
         Args:
@@ -475,7 +284,7 @@ class ChunkPersister:
         collection_metadata: dict = {},
         save_metadata=False,
         use_memorycache=True,
-        cache = None
+        cache=None,
     ):
         """Chunks every incoming dekriptor into subchunks if deskriptor is larger than segment_slice
          or extends the deskriptor to the respective chunksize if deskriptor is smaller than segment_slice
@@ -572,7 +381,7 @@ class ChunkPersister:
 
     def configure(self, deskriptor=None):
         rs = deskriptor["self"]
-        if rs.get("bypass",False):
+        if rs.get("bypass", False):
             return deskriptor
 
         def get_value(attr_name):
@@ -622,7 +431,7 @@ class ChunkPersister:
                 global_lock=self.mutex,
                 save_metadata=self.save_metadata,
                 cache=self.cache,
-                use_memorycache=self.use_memorycache
+                use_memorycache=self.use_memorycache,
             )
             cloned_persister.dask_key_name = self.dask_key_name + "_persister"
             dict_update(
@@ -646,7 +455,9 @@ class ChunkPersister:
 
         return deskriptor
 
-    def compute(self, *data: DataContainer | NodeFailedException, **deskriptor) -> DataContainer:
+    def compute(
+        self, *data: DataContainer | NodeFailedException, **deskriptor
+    ) -> DataContainer:
         def unpack_list(inputlist):
             new_list = []
             for item in inputlist:
